+++ sources/www/widgets/gauge/viewModel.js
@@ -66,8 +66,7 @@ widgetViewModelCtor = function gaugeViewModel() {
                }
            }
         }
-        catch(error)
-        {
+        catch(error){
            self.widgetApi.setState (widgetStateEnum.InvalidConfiguration);
         }
     };
@@ -79,7 +78,10 @@ widgetViewModelCtor = function gaugeViewModel() {
             return;
 
         //we register keyword new acquisition
-        self.widgetApi.registerKeywordAcquisitions(self.widget.configuration.device.keywordId);
+        self.widgetApi.registerKeywordForNewAcquisitions(self.widget.configuration.device.keywordId);	   
+	   
+        //we register keyword for get last value at web client startup 
+        self.widgetApi.getLastValue(self.widget.configuration.device.keywordId); 
         
         //we fill the deviceId of the battery indicator
         self.widgetApi.configureBatteryIcon(self.widget.configuration.device.deviceId);