+++ sources/www/widgets/moon/viewModel.js
@@ -60,6 +60,9 @@ widgetViewModelCtor = function moonViewModel() {
 
     this.configurationChanged = function () {
         //we register keyword new acquisition
-        this.widgetApi.registerKeywordAcquisitions(this.widget.configuration.device.keywordId);
+        self.widgetApi.registerKeywordForNewAcquisitions(self.widget.configuration.device.keywordId);	   
+	   
+		//we register keyword for get last value at web client startup
+		self.widgetApi.getLastValue(self.widget.configuration.device.keywordId); 
     };
 };