+++ sources/www/widgets/light/changelog.md
@@ -1,2 +1,5 @@
+### 1.0.4
+* Remove obsolete function
+
 ### 1.0.3
 * Better display