+++ sources/www/widgets/shutter/changelog.md
@@ -1,2 +1,5 @@
+### 1.0.4
+* Remove obsolete function
+
 ### 1.0.3
 * fix return default value, when "verticalInverter". Avoid a console error.