+++ sources/www/widgets/thermometer/changelog.md
@@ -1,3 +1,7 @@
+### 1.0.2
+* Remove obsolete function
+
+
 ### 1.0.1
 * Fix scale thermometer.
 * Fix divide by 0